requirements.txt
moviepy==1.0.3
requests
google-genai
Pillow
python-dotenv
elevenlabs